refactor(HealthReportsScreen): mover estilos para arquivo styles.js
